+++ src/main/java/domain/Player.java
@@ -4,40 +4,18 @@ import domain.card.Card;
 
 import java.util.List;
 
-public class Player {
-    private final String name;
-    private final Hand hand;
+public class Player extends Participant {
 
     public Player(String name) {
-        this.name = name;
-        hand = new Hand();
+        super(name);
     }
 
     public void draw(Card card) {
         hand.addCard(card);
     }
 
-    public boolean isBurst() {
-        return hand.isBurst();
-    }
-
+    @Override
     public boolean canHit() {
         return hand.isLessThanBlackJack();
-    }
-
-    public int getScore() {
-        return hand.getSum();
-    }
-
-    public List<Card> getCards() {
-        return hand.getCards();
-    }
-
-    public int getHandSize() {
-        return hand.getSize();
-    }
-
-    public String getName() {
-        return name;
     }
 }
